Rubocopped changes to observation report.
